Move Skill class into its own separate sources
Isolate the Skill class into its own set of source files in preparation for
harnessing it in a unit test as a first step to replacing CMarkup with TinyXML2
and eventual refactoring into a moddable target.
